# api/openapi.yaml
openapi: 3.1.0
info:
  title: scanX API
  version: 1.0.0-phase2
  description: |
    scanX REST API (spec §9). Authenticate with `Authorization: Bearer scanx_pat_…`
    (personal API token) or with the browser session cookie plus `X-CSRF-Token`.
    Resources of organizations you do not belong to are reported as 404.
  license:
    name: BUSL-1.1
    url: https://github.com/ininia/scanx/blob/main/LICENSE
servers:
  - url: /api/v1
security:
  - bearer: []
  - session: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Log in with e-mail and password (sets the session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Logged in; if mfa_required, POST /auth/mfa next
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /auth/mfa:
    post:
      tags: [auth]
      security: []
      summary: Complete a 2FA login with a TOTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      tags: [auth]
      responses:
        "204": { description: Logged out }
        "401": { $ref: "#/components/responses/Error" }
  /me:
    get:
      tags: [auth]
      responses:
        "200":
          description: The caller
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Error" }
  /orgs:
    get:
      tags: [orgs]
      responses:
        "200":
          description: Organizations the caller belongs to
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrgPage" }
        "401": { $ref: "#/components/responses/Error" }
    post:
      tags: [orgs]
      summary: Create an organization (superadmin, or signup mode)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                slug: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Org" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /orgs/{org}:
    parameters: [{ $ref: "#/components/parameters/org" }]
    get:
      tags: [orgs]
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Org" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [orgs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Org" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /orgs/{org}/members:
    parameters: [{ $ref: "#/components/parameters/org" }]
    get:
      tags: [members]
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberPage" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [members]
      summary: Add an existing user by e-mail
      requestBody: { $ref: "#/components/requestBodies/EmailRole" }
      responses:
        "204": { description: Added }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /orgs/{org}/members/{user}:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { name: user, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "204": { description: Updated }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      tags: [members]
      responses:
        "204": { description: Removed }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /orgs/{org}/invitations:
    parameters: [{ $ref: "#/components/parameters/org" }]
    get:
      tags: [members]
      responses:
        "200":
          description: Open invitations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPage" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [members]
      summary: Invite by e-mail; the response contains the one-time token
      requestBody: { $ref: "#/components/requestBodies/EmailRole" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /orgs/{org}/invitations/{id}:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/id" }
    delete:
      tags: [members]
      responses:
        "204": { description: Revoked }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/tokens:
    parameters: [{ $ref: "#/components/parameters/org" }]
    get:
      tags: [tokens]
      responses:
        "200":
          description: The caller's API tokens in this organization
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPage" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [tokens]
      summary: Create a personal API token (shown once)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string }
                scopes: { type: array, items: { type: string, enum: [read, write] } }
                expires_in_days: { type: integer, minimum: 0, maximum: 3650 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Token" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /orgs/{org}/tokens/{id}:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/id" }
    delete:
      tags: [tokens]
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/audit:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/page" }
      - { $ref: "#/components/parameters/per_page" }
    get:
      tags: [orgs]
      responses:
        "200":
          description: Audit log (newest first); total is -1 (not counted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditPage" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/projects:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/page" }
      - { $ref: "#/components/parameters/per_page" }
    get:
      tags: [projects]
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectPage" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [projects]
      requestBody: { $ref: "#/components/requestBodies/Project" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /orgs/{org}/projects/{project}:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { name: project, in: path, required: true, schema: { type: string } }
    get:
      tags: [projects]
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [projects]
      requestBody: { $ref: "#/components/requestBodies/Project" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      tags: [projects]
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/projects/{project}/deploy-key:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/project" }
    get:
      tags: [scans]
      summary: Public half of the read-only deploy key of the project
      responses:
        "200":
          description: Deploy key
          content:
            application/json:
              schema:
                type: object
                required: [public_key, fingerprint]
                properties:
                  public_key: { type: string }
                  fingerprint: { type: string }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/projects/{project}/scans:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/project" }
    get:
      tags: [scans]
      responses:
        "200":
          description: Recent scans (newest first)
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Scan" } }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [scans]
      summary: Queue a scan (default branch when omitted)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                branch: { type: string }
      responses:
        "202":
          description: Queued (or the already queued scan of this branch)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scan" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /orgs/{org}/projects/{project}/issues:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/project" }
      - { name: status, in: query, required: false, schema: { type: string, enum: ["", open, fixed, false_positive, accepted_risk, wont_fix] } }
    get:
      tags: [issues]
      responses:
        "200":
          description: Issues (max 500)
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Issue" } }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/scans/{scan}:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/scan" }
    get:
      tags: [scans]
      responses:
        "200":
          description: Scan with tool runs
          content:
            application/json:
              schema:
                type: object
                required: [scan, project, tools]
                properties:
                  scan: { $ref: "#/components/schemas/Scan" }
                  project: { type: string }
                  tools: { type: array, items: { type: object } }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/scans/{scan}/issues:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/scan" }
    get:
      tags: [scans]
      responses:
        "200":
          description: Issues found by this scan
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Issue" } }
        "404": { $ref: "#/components/responses/Error" }
  /orgs/{org}/scans/{scan}/cancel:
    parameters:
      - { $ref: "#/components/parameters/org" }
      - { $ref: "#/components/parameters/scan" }
    post:
      tags: [scans]
      responses:
        "204": { description: Canceled }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    session: { type: apiKey, in: cookie, name: __Host-scanx_session }
  parameters:
    org: { name: org, in: path, required: true, schema: { type: string } }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    LoginResult:
      type: object
      required: [mfa_required]
      properties:
        mfa_required: { type: boolean }
    Me:
      type: object
      required: [id, email, name, is_superadmin, locale, via_token]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
        is_superadmin: { type: boolean }
        locale: { type: string }
        via_token: { type: boolean }
    Org:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        created_at: { type: string, format: date-time }
...
    OrgPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - properties: { items: { type: array, items: { $ref: "#/components/schemas/Org" } } }
    MemberPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - properties: { items: { type: array, items: { $ref: "#/components/schemas/Member" } } }
...